% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_dissimilarity.R
\name{calc_dissimilarity}
\alias{calc_dissimilarity}
\title{Invoke DIAMOND to calculate alignment-based similarity scores}
\usage{
calc_dissimilarity(
  proteins,
  min_align = 6,
  diamond_folder = "./",
  save_folder = "./",
  verbose = TRUE,
  mop_up = FALSE
)
}
\arguments{
\item{proteins}{data frame of protein sequence data. Must have at least
two columns, \code{UID} (with the protein IDs) and \code{TSeq_sequence} (with the
protein sequences). This data frame is usually created using
\code{\link[=get_proteins]{get_proteins()}}.}

\item{min_align}{smallest alignment size to consider.}

\item{diamond_folder}{path to folder where DIAMOND can be run. Defaults to
the current working directory (which works, e.g., when a system-wide install
is present).}

\item{save_folder}{path to folder for saving the results. Use \code{NULL}
if saving to file is not desired. Defaults to
the current working directory.}

\item{mop_up}{logical: should the DIAMOND files be deleted upon completion?}
}
\value{
Dissimilarity matrix for the proteins queried.
}
\description{
This function calculates the dissimilarity score between protein sequences
in data frame \code{proteins}. NOTE: this function requires
DIAMOND to be installed. Depending on the OS, this may mean a proper
system-wide installation or simply the presence of the executable in the
appropriate folder. Please check
\url{https://github.com/bbuchfink/diamond/wiki/2.-Installation}
for details.
}
\section{Dissimilarity calculation}{

Dissimilarity is calculated based on the local alignment scores returned by
DIAMOND. Alignments shorter than \code{min_align} are ignored. The dissimilarity
between a pair of protein sequences is calculated as $1 - max(pident)/100$,
where \code{pident} is the vector of local alignment scores returned for that
pair of proteins. When no alignment is found for a given pair, the
dissimilarity is set to 1.
}

\author{
Felipe Campelo (\email{f.campelo@aston.ac.uk})
}
