% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_features_ESM.R
\name{calc_ESM_features}
\alias{calc_ESM_features}
\title{Calculate classic features for epitope prediction}
\usage{
calc_ESM_features(
  prot.df,
  py_script_path,
  id_column = "Info_protein_id",
  seqs_column = "Info_protein_sequence",
  save_folder = "./esm1b_features",
  model = "esm1b_t33_650M_UR50S",
  model.params = "--include per_tok --nogpu --repr_layers 33",
  chunk_size = 1000,
  step_size = 50,
  ncpus = 1,
  feat_prefix = "feat_esm1b_"
)
}
\arguments{
\item{prot.df}{dataframe containing a column with proteins
sequences}

\item{py_script_path}{path to the ESM Python script (NOTE: this is
provided as inst/utils/extract.py in this package's folder structure. The
script is a slightly modified version of the one available in
\url{https://github.com/facebookresearch/esm/blob/main/scripts/extract.py}).}

\item{id_column}{name of column in \code{prot.df} containing the unique protein
IDs.}

\item{seqs_column}{name of column in \code{prot.df} containing the protein
sequences}

\item{save_folder}{path to folder for saving the output.}

\item{model}{string with the full model name
(see \url{https://github.com/facebookresearch/esm})}

\item{model.params}{string with model options to be used}

\item{chunk_size}{size of chunk to be used when proteins are
longer than 1024 residues (usually as large as possible and < 1024)}

\item{step_size}{step size to use for processing long proteins. Smaller
values are better, but more computationally intensive.}

\item{ncpus}{positive integer, number of cores to use}

\item{feat_prefix}{prefix to be added to the feature names.}
}
\description{
This function is used to calculate features using the ESM embedders,
particularly ESM-1b. It is a \strong{very} computationally demanding model,
and usually cannot run in regular computers (this function is intended
for use in high-performance workstations or clusters). The function also
needs a good internet connection to run, as the ESM model requires the
download of certain elements to be run.
To check the details and install instructions of ESM, please see
\url{https://github.com/facebookresearch/esm} and examples therein. A clone
of that repository is also available at \url{https://github.com/fcampelo/esm}.
}
\section{Sequence length and non-standard AAs}{

ESM-1b returns errors when trying to calculate features for proteins longer
than 1024 residues (see https://github.com/facebookresearch/esm/issues/49).
A possible solution (suggested in, e.g.,
https://github.com/brianhie/evolocity/issues/2)
is to break longer sequences in smaller overlapping windows, predict
independently and then aggregate in post-processing, averaging where needed.
We follow this strategy here.

For the feature calculation, any non-standard AA character is replaced by
the \verb{<mask>} placeholder.
}

