% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_features_classic.R
\name{calc_features_classic}
\alias{calc_features_classic}
\title{Calculate classic features for epitope prediction}
\usage{
calc_features_classic(
  X,
  seqs_column,
  features,
  cl = NULL,
  calc_extended = FALSE
)
}
\arguments{
\item{X}{a peptide.list object (returned by \code{\link[=extract_labelled_data]{extract_labelled_data()}} or a
data frame containing one column with sequences for feature
calculation.}

\item{seqs_column}{name of the column containing the sequences}

\item{features}{vector of names of features to be calculated.
See \strong{Feature Vector} for details.}

\item{cl}{a SOCK cluster object created using \code{\link[=set_mc_cluster]{set_mc_cluster()}}, or
\code{NULL} if parallel processing is not desired.}

\item{calc_extended}{logical: should the features be calculated for
the extended dataframe (i.e., containing all positions and not only the
labelled ones), if available?}
}
\value{
Updated \code{X} with features appended as columns
(directly to \code{X} if it is a data.frame, or to \code{X$df} if it is a peptide.list object)
}
\description{
This function is used to calculate several distinct families of features for
epitope prediction. These are 'classic' features, based on statistical
summaries and average physico-chemical characteristics of the local
neighborhood of each residue in a protein sequence, or of the protein as a
whole.
}
\details{
The following features are calculated based on the implementations available
in package
\href{https://cran.r-project.org/package=protr}{\strong{protr}}. As of December 2023,
the following groups of features are supported by the \strong{epitopes} package
(see the \href{https://CRAN.R-project.org/package=protr/vignettes/protr.html#3_package_overview}{protr vignette}
for details on each of these):
\itemize{
\item "AAC" - Amino acid composition
\item "DC"  - Dipeptide composition
\item "TC"  - Tripeptide composition
\item  CTD descriptors:
\itemize{
\item "CTDC" - Composition
\item "CTDT" - Transition
\item "CTDD" - Distribution
}
\item "CTriad" - Conjoint triad descriptors
\item  Quasi-sequence-order descriptors:
\itemize{
\item "SOCN" - Sequence-order-coupling number (with maximum lag
\code{nlag = 3})
\item "QSO" - Quasi-sequence-order descriptors (with maximum lag
\code{nlag = 3} and weighting factor \code{w = 0.1})
}
\item Proteochemometric Modeling descriptors:
\itemize{
\item "ScalesGap" - Scales-based descriptors derived by Principal
Components Analysis (using all properties in the \code{protr::AAindex}
matrix, \code{pc = 5} and \code{lag = 3})
}
}

\strong{NOTE}: in all feature groups except "ScalesGap", invalid AA codes
(B, J, O, U, X, Y) are removed from the strings prior to feature calculation.
In "ScalesGap" these codes are replaced by a gap indicator, "-".

Besides those, the following features are also available based on native
implementations:
\itemize{
\item "Entropy" - the Shannon entropy of a sequence.
\item "Atoms" - the number of C, H, N, O, S atoms in the sequence
\item "MolWeight" - the total molecular weight of the peptide
\item "AAtypes" - the proportion of AAs of each type (acidic, aliphatic,
acidic, etc.)
\item "LegacyFeatures" - calculates the features used in paper
\doi{10.1093/bioinformatics/btab536}
}

Each feature group may be used for peptides or full protein sequences
Note, however, the warning from the \strong{protr} documentation:
"\emph{Users need to intelligently evaluate the underlying}
\emph{details of the descriptors provided, instead of using protr with their data}
\emph{blindly, especially for the descriptor types with more flexibility. It}
\emph{would be wise for the users to use some negative and positive control}
\emph{comparisons where relevant to help guide interpretation of the results.}".
Users should therefore be savvy when choosing which features to use for
epitope prediction, and the choice should ideally be guided by domain
expertise. Certain feature groups may not make sense for short peptides,
as they may be almost completely uninformative (e.g., tripeptide composition,
"TC"); or may require specific overriding of standard parameters (e.g.,
parameter \code{lambda} for "PAAC" must be smaller than the length of the shortest
peptide).
}
\section{\strong{Feature Vectors}}{

Input vector \code{features} is used to define which features are calculated
for proteins or peptides. This input parameter must be a
character vector, where each element is one of the names listed in
\strong{Details}. For more information on the features calculated by \strong{protr},
check \code{?protr::extractXYZ}, replacing \code{XYZ} by the group abbreviation (see
\strong{Details} or the documentation of the \strong{protr} package for the list of
available feature groups). Notice that, for consistency purposes, the user
parameters of all \strong{protr} features are kept fixed in this routine.
If the user wishes to add other features (or the same features with distinct
parameters) they can calculate those separately
and bind them to \code{X} (or to \code{X$df}, if X is a peptide list).
All feature columns calculated in this function
will have names starting with "feat_", and if externally-calculated features
are added they should follow the same pattern (for compatibility with the
other functions in this package).
}

\author{
Felipe Campelo (\email{fcampelo@gmail.com})
}
