% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_features_esm2.R
\name{calc_features_esm2}
\alias{calc_features_esm2}
\title{Invoke model to calculate ESM2 features}
\usage{
calc_features_esm2(
  X,
  mode = "run",
  seqs_column = "Info_protein_sequence",
  ids_column = "Info_protein_id",
  cl = NULL,
  save_folder = "./data",
  model_spec = "esm2_t33_650M_UR50D",
  model_opts = "--include per_tok --repr_layers 33",
  delete_protein_csv_files = FALSE
)
}
\arguments{
\item{cl}{a SOCK cluster object created using \code{\link[=set_mc_cluster]{set_mc_cluster()}}, or
\code{NULL} if parallel processing is not desired.}
}
\description{
Invoke model to calculate ESM2 features
}
\author{
Felipe Campelo (\email{fcampelo@gmail.com})
}
