% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concatenate_esm_outputs.R
\name{concatenate_esm_outputs}
\alias{concatenate_esm_outputs}
\title{Consolidate the outputs from calc_features_esm2}
\usage{
concatenate_esm_outputs(
  csv_folder,
  save_folder = NULL,
  filenames = NULL,
  save_file = NULL,
  feat_prefix = "feat_esm2_",
  cl = NULL,
  delete_originals = FALSE
)
}
\arguments{
\item{cl}{a SOCK cluster object created using \code{\link[=set_mc_cluster]{set_mc_cluster()}}, or
\code{NULL} if parallel processing is not desired.}
}
\description{
Consolidate the outputs from calc_features_esm2
}
\author{
Felipe Campelo (\email{fcampelo@gmail.com})
}
