% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_labelled_data.R
\name{extract_labelled_data}
\alias{extract_labelled_data}
\title{Extract peptides and windowed representation for epitope prediction.}
\usage{
extract_labelled_data(
  df,
  min_peptide = 8,
  max_epitope = 30,
  window_size = (2 * min_peptide) - 1,
  clean_main_df = TRUE,
  save_folder = NULL
)
}
\arguments{
\item{df}{data frame of consolidated protein-epitope data, returned by
\code{\link[=consolidate_data]{consolidate_data()}}.}

\item{min_peptide}{positive integer, shortest peptide to be considered.
Peptides shorter than \code{min_peptide} have their class label reverted to \code{NA} to
prevent very short entries from adding too much noise into the data.}

\item{max_epitope}{positive integer, longest positively-labelled peptide to
be considered. Positively-labelled peptides longer than \code{max_epitope} have
their class label reverted to \code{NA} to prevent long epitope-containing regions from
adding too much noise into the data.}

\item{window_size}{positive integer, size of the local neighbourhood to be
considered.}

\item{clean_main_df}{flag: should the main dataframe (\verb{$df} in the output list)
be filtered to remove positions without class labels or that don't comply with
\code{min_peptide} and \code{max_epitope}?}

\item{save_folder}{path to folder for saving the results. It will save the
results as file \emph{peptides_list.rds} (overwriting if necessary)}
}
\value{
List containing three data frames:

\itemize{
\item \strong{df}:  data frame containing the labeled positions of \code{df}
(filtered by \emph{min_peptide} and \emph{max_epitope}) with a new column
\strong{Info_window} containing the local neighbourhood of each position
(to be used for feature calculation)
\item \strong{peptides}: data frame containing one peptides per row, together
with its calculated Class attribute value.
\item \strong{proteins}: data frame containing one protein per row, with all
proteins that appear at least once in \code{df} and \code{peptides}.
}
}
\description{
Extract relevant peptides from the consolidated protein-epitope data
(generated using \code{\link[=consolidate_data]{consolidate_data()}}) and build column with neighbourhood
of each labeled aminoacid residue.
}
\author{
Felipe Campelo (\email{fcampelo@gmail.com})
}
