% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_model.R
\name{fit_model}
\alias{fit_model}
\title{Fit Random Forest model to epitope data}
\usage{
fit_model(
  peptides.list,
  threshold = 0.5,
  holdout.split = NULL,
  CV.folds = NULL,
  sample.rebalancing = TRUE,
  use.global.features = ifelse(peptides.list$splits.attrs$split_level == "protein",
    TRUE, FALSE),
  ncpus = 1,
  rnd.seed = NULL,
  ...
)
}
\arguments{
\item{peptides.list}{data frame containing the training data (one or more
numerical predictors and one \strong{Class} attribute).}

\item{threshold}{probability threshold for attributing a prediction as
\emph{positive}.}

\item{holdout.split}{name of split to be used as a holdout set. If \code{NULL}
then the full data is used for training.}

\item{CV.folds}{vector with the names of the splits to be used as CV folds.
If \code{NULL} no cross-validation is performed.}

\item{sample.rebalancing}{logical: should the model try to compensate class
imbalances by weighted sampling of examples when training the trees in the
random forest? See \strong{Dealing with class imbalance}.}

\item{use.global.features}{logical: should global features (potentially
available in \code{peptides.list$proteins}) be used? Should be left as the default
unless the user knows exactly what they're doing. See \strong{Details}.}

\item{ncpus}{number of cores to use.}

\item{rnd.seed}{seed for random number generator. \strong{Note}: this function
always returns the state of the random number generator back to its original
value before returning the results. Running it twice in sequence with
\code{rnd.seed = NULL} should result in exactly the same results (but it is
safer to simply set a specific seed, or to reuse the seed that is returned
in the output list).}

\item{...}{other options to be passed down to \code{\link[ranger:ranger]{ranger::ranger()}}.}
}
\value{
List containing the fitted model and several performance indicators.
}
\description{
Fits a Random Forest model to epitope data, using the data splits and
features previously calculated using \code{\link[=make_data_splits]{make_data_splits()}} and
\code{\link[=calc_features]{calc_features()}}, respectively.
}
\details{
Function \code{\link[=make_data_splits]{make_data_splits()}} defines data splits based on (protein or
peptide) similarity. The split identifiers are stored in
\code{peptides.list$df$Info_split}. Function \code{\link[=calc_features]{calc_features()}} calculates
the local and/or global features for each entry. Local features are stored in
\code{peptides.list$df} as columns starting with \code{feat_local_}, whereas global
features are stored in \code{peptides.list$proteins}, as columns starting with
\code{feat_global_}.
\strong{NOTE}: Global features should only be used if the splitting level
used was "protein", otherwise they may cause contamination of performance
assessment due to data leakage. The splitting level of the data can be
checked on \code{peptides.list$splits.attrs$split_level}.
}
\section{Dealing with class imbalance}{

Parameter \code{sample.rebalancing} regulates whether the resulting model attempts
to compensate class imbalances. If \code{TRUE} the Random Forest model is subject
to cost-sensitive training, which is done internally by setting the
parameter \code{case.weights} in the call to \code{\link[ranger:ranger]{ranger::ranger()}} to a vector where
each observation of class \emph{i} has a weight equal to \code{1 / K_i}, where
\code{K_i} is the total number of cases of class \code{i} in the training data.
}

\section{Performance assessment}{

This function has the following modes of performance assessment:

\itemize{
\item If both \emph{holdout.split} and \emph{CV.folds} are \code{NULL}, no performance
assessment is done. A model is fit on the full data and returned.
\item If \emph{holdout.split} is the valid name of a data split in
\code{peptides.list$df$Info_split} \strong{AND} \emph{CV.folds} is \code{NULL}, a model is
trained on the full data except the \emph{holdout.split} and then applied to
predict the labels for \emph{holdout.split}. The performance returned
corresponds to the performance on \emph{holdout.split}.
\item If \emph{CV.folds} is a vector of valid names of data splits in
\code{peptides.list$df$Info_split}, the splits named are used as
cross-validation folds. The performance returned
corresponds to the average cross-validation performance using the
\emph{CV.folds}. Notice that if \emph{CV.folds} is not \code{NULL} the value of
\emph{holdout.split} is ignored.
}

\strong{IMPORTANT}: in all cases, the model returned by this routine is a model
trained on the \strong{full data} (fit after the performance is assessed on
holdout splits or on cross-validation folds).
}

\author{
Felipe Campelo (\email{f.campelo@aston.ac.uk})
}
