% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_RF.R
\name{fit_RF}
\alias{fit_RF}
\title{Fits a single random forest and assess the performance on a holdout split.}
\usage{
fit_RF(
  df,
  sample.rebalancing,
  holdout.split = NULL,
  threshold,
  ncpus,
  min.node.size,
  ntrees,
  ...
)
}
\description{
Not to be used separately, this function is called as part of calls to
\code{\link[=fit_model]{fit_model()}}.
}
\details{
@inheritParams fit_model
@param df dataframe containing a \code{Class} attribute and one or more feature
attributes (columns starting with "feat_")
@param min.node.size min node size for \code{\link[ranger:ranger]{ranger::ranger()}}
@param ntrees number of trees for \code{\link[ranger:ranger]{ranger::ranger()}}

@importFrom dplyr \%>\%
@importFrom rlang .data
@export
}
