% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_basic_data.R
\name{get_basic_data}
\alias{get_basic_data}
\title{Retrieve basic pre-computed datasets for linear B-cell epitope prediction}
\usage{
get_basic_data(
  save_folder,
  datasets = c("epitopes", "proteins", "taxonomy", "protein_dissimilarity"),
  timeout = 60
)
}
\arguments{
\item{save_folder}{path to folder for saving the results.}

\item{datasets}{character vector with the dataset(s) to be retrieved. Accepts
any combination of "epitopes", "proteins", "taxonomy" or "protein_dissimilarity".}

\item{timeout}{maximum download time per dataset.}
}
\value{
This function returns \code{TRUE} (invisibly) upon completion.
}
\description{
This function can be used to retrieve pre-computed base data sets
extracted from the Immune Epitope DataBase (IEDB). The datasets are:
\itemize{
\item epitopes.rds (containing consolidated LBCE data)
\item proteins.rds (containing all proteins related to epitopes.rds)
\item taxonomy.rds (containing all taxonomic relations from the organisms listed
in epitopes.rds)
\item protein_dissimilarity.rds (normalised local dissimilarity matrix for all
proteins in proteins.rds, calculated using \href{https://github.com/bbuchfink/diamond}{Diamond}).
}
}
\author{
Felipe Campelo (\email{f.campelo@aston.ac.uk})
}
