% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_esm_features.R
\name{get_esm_features}
\alias{get_esm_features}
\title{Retrieve pre-calculated ESM-1b protein features.}
\usage{
get_esm_features(
  path,
  Info_protein_id = NULL,
  Info_organism_id = NULL,
  save_folder = NULL,
  cl = cl
)
}
\arguments{
\item{path}{string containing the path to the folder where the feature
files are located. There must be a tab-separated file in this folder called
\code{filemap.tsv} with columns \code{Info_protein_id} (protein ids), \code{folder}
(relative subfolder containing that protein) and \code{Info_organism_id}
(comma-separated string with the taxonomy IDs of pathogens associated with
the protein in the IEDB records).}

\item{Info_protein_id}{vector of protein IDs.}

\item{Info_organism_id}{vector of organism taxonomy IDs.}

\item{save_folder}{path to folder for saving the results.}

\item{cl}{a SOCK cluster object created using \code{\link[=set_mc_cluster]{set_mc_cluster()}}, or
\code{NULL} if parallel processing is not desired.}
}
\value{
A named list vector containing the data frames with the ESM features
for each protein. IDs that are not available in \code{path} are returned as
empty list positions. This list vector has attributes \code{OK} (vector of
ids of proteins successfully retrieved) and \code{FAIL} (unsuccessful ids).
}
\description{
This function retrieves pre-calculated ESM-1b features given a
vector of protein IDs or a vector of taxonomy IDs.
}
\author{
Felipe Campelo (\email{fcampelo@gmail.com})
}
