% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_linear_epitopes.R
\name{get_LBCE}
\alias{get_LBCE}
\title{Extract LBCE data from IEDB export}
\usage{
get_LBCE(data_folder, cl = NULL, save_folder = NULL)
}
\arguments{
\item{data_folder}{path (either relative or absolute) to the directory
containing the XML files}

\item{cl}{a SOCK cluster object created using \code{\link[=set_mc_cluster]{set_mc_cluster()}}, or
\code{NULL} if parallel processing is not desired.}

\item{save_folder}{path to folder for saving the output.}
}
\description{
Alias function for \code{\link[=get_linear_epitopes]{get_linear_epitopes()}} with type = "B"
}
