% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_linear_epitopes.R
\name{get_LTCE}
\alias{get_LTCE}
\title{Exgtract LTCE data from IEDB export}
\usage{
get_LTCE(data_folder, ncpus = 1, save_folder = NULL)
}
\arguments{
\item{data_folder}{path (either relative or absolute) to the directory
containing the XML files}

\item{ncpus}{positive integer, number of cores to use}

\item{save_folder}{path to folder for saving the output.}
}
\description{
Alias function for \code{\link[=get_linear_epitopes]{get_linear_epitopes()}} with type = "T"
}
