% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_precalc_features.R
\name{get_precalc_features}
\alias{get_precalc_features}
\title{Retrieve protein data with pre-computed features}
\usage{
get_precalc_features(proteinIDs, model = "esm1b", save_folder, ntries = 3)
}
\arguments{
\item{proteinIDs}{A character vector with protein IDs.}

\item{model}{which embedding to retrieve. Currently accepts "esm1b".}

\item{save_folder}{path to folder for saving the results.}

\item{ntries, }{number of times the routine will try to retrieve each dataset
before giving up}
}
\value{
Vector of retrieval status
}
\description{
This function can be used to retrieve pre-computed features for proteins
containing at least one peptide labelled as a linear B-cell epitope (or
non-LBCE) in IEDB. The features are calculated using the following feature embeddings:
}
\details{
\itemize{
\item ESM-1b embedder
(\url{https://github.com/facebookresearch/esm})
}
}
\author{
Felipe Campelo (\email{f.campelo@aston.ac.uk})
}
