% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_precomputed_datasets.R
\name{get_precomputed_datasets}
\alias{get_precomputed_datasets}
\title{Retrieve linear B-cell data with ESM-1b features}
\usage{
get_precomputed_datasets(
  taxonIDs,
  tax_list,
  ntries = 3,
  min_peptide = 5,
  max_epitope = 30,
  save_folder = NULL
)
}
\arguments{
\item{taxonIDs}{A numeric or character vector with taxonomy IDs.
All LBCE entries related to organisms under the taxonomy IDs listed
will be returned.}

\item{tax_list}{taxonomy list (generated by \code{\link[=get_taxonomy]{get_taxonomy()}} or retrieved
using \code{\link[=get_basic_data]{get_basic_data()}}).
#' @param min_peptide positive integer, shortest peptide to be considered
(applies to both \strong{positive} and \strong{negative} observations)}

\item{ntries, }{number of times the routine will try to retrieve each dataset
before giving up}

\item{max_epitope}{positive integer, longest peptide to be considered (only
applies to \strong{positive} observations.)}

\item{save_folder}{path to folder for saving the results.}
}
\value{
Returns list of retrieved datasets (invisibly).
}
\description{
This function can be used to retrieve pre-computed LBCE datasets
extracted from the Immune Epitope DataBase (IEDB). The datasets
contain the features calculated using the ESM-1b embedder (
\url{https://github.com/facebookresearch/esm})
}
\author{
Felipe Campelo (\email{f.campelo@aston.ac.uk})
}
