% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_precomputed_proteins.R
\name{get_precomputed_proteins}
\alias{get_precomputed_proteins}
\title{Retrieve protein data with ESM-1b features}
\usage{
get_precomputed_proteins(proteinIDs, save_folder, ntries = 3)
}
\arguments{
\item{proteinIDs}{A character vector with protein IDs.}

\item{save_folder}{path to folder for saving the results.}

\item{ntries, }{number of times the routine will try to retrieve each dataset
before giving up}
}
\value{
Vector of retrieval status
}
\description{
This function can be used to retrieve pre-computed features for proteins
containing at least one peptide labelled as a linear B-cell epitope (or
non-LBCE) in IEDB. The features are calculated using the ESM-1b embedder
(\url{https://github.com/facebookresearch/esm})
}
\author{
Felipe Campelo (\email{f.campelo@aston.ac.uk})
}
