% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_proteins.R
\name{get_proteins}
\alias{get_proteins}
\title{Retrieve protein sequences and data from GenBank and Uniprot}
\usage{
get_proteins(uids, save_folder = NULL)
}
\arguments{
\item{uids}{A character vector with protein IDs.}

\item{save_folder}{path to folder for saving the results.}
}
\value{
A data frame with the extracted proteins.
}
\description{
This function is used to retrieve data from protein databases. IDs are checked on NCBI and UniprotKB.
}
\details{
Queries are processed one by one (rather than in batch) to enable treatment
of individual inconsistencies (e.g., wrong UIDs, queries that return a
different identifier, etc.). This makes this routine substantially slower,
but considerably more robust to errors.
}
\author{
Felipe Campelo (\email{f.campelo@aston.ac.uk})
}
