% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_bioc_dependencies.R
\name{install_bioc_dependencies}
\alias{install_bioc_dependencies}
\title{Install Bioconductor dependencies}
\usage{
install_bioc_dependencies(bioc.args = list(), force = FALSE)
}
\arguments{
\item{bioc.args}{list containing further arguments to be passed
down to \code{BiocManager::install()}.}

\item{force}{logical: reinstall already-installed packages?}
}
\value{
No return value, called for side effects.
}
\description{
This function installs the latest versions of all Bioconductor dependencies
of the \code{epitopes} package, namely:
\itemize{
\item Biostrings
}
}
\details{
It is essential that the package(s) above be installed for \code{epitopes} to
work properly.
}
\examples{
\dontrun{
  install_bioc_dependencies()
}

}
