% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_data_splits.R
\name{make_data_splits}
\alias{make_data_splits}
\title{Define data splits for machine learning.}
\usage{
make_data_splits(
  peptides.list,
  delta,
  w = c(0.5, 0.4, 0.1),
  split_names = NULL,
  similarity_threshold = 0.7,
  target_id = NULL,
  tax_list = NULL,
  save_folder = NULL,
  ncpus = 1,
  cdhit.par.list = list(s = 0.5)
)
}
\arguments{
\item{peptides.list}{list object returned by \code{\link[=extract_labelled_data]{extract_labelled_data()}}, containing
the data frame of windowed epitope data (\strong{$df}), the data frame of
individual peptides (\strong{$peptides}) and the data frame of proteins
(\strong{$proteins}).}

\item{delta}{vector of desired split proportions (must add up to one). It
is useful (but not mandatory) to use a vector that is sorted
in decreasing order.}

\item{w}{vector of weights for function aggregation. Must be of
length 3 and add to 1. All weights must be non-negative. See
package \code{moses} for details.
\code{w[1]}: weight of the objective to obey desired split proportions;
\code{w[2]}: weight of the objective to distribute classes proportionally between splits;
\code{w[3]}: weight of the objective to maximise within-splits diversity.}

\item{split_names}{optional, vector of names to be given to each split.}

\item{similarity_threshold}{similarity threshold for grouping observations.
\eqn{1 - similarity_threshold} is passed to \code{\link[CellaRepertorium:cdhit]{CellaRepertorium::cdhit()}} as}

\item{target_id}{vector of lower-level taxonomy ids (integer or character)
that the splitting strategy should try to prioritise
distributing across splits (useful, e.g., if generating splits at a
higher taxonomic level for later re-use at a lower level).}

\item{tax_list}{taxonomy list (generated by \code{\link[=get_taxonomy]{get_taxonomy()}}). Only used if
\code{target_id} is not \code{NULL}.}

\item{save_folder}{path to folder for saving the results. It will save the
results as file \emph{peptides_list.rds} (overwriting if necessary)}

\item{ncpus}{positive integer, number of cores to use.}

\item{cdhit.par.list}{further parameters to be passed down to
\code{\link[CellaRepertorium:cdhit]{CellaRepertorium::cdhit()}}. Please check that function for details.}
}
\value{
Updated \code{peptides.list} with dataframes containing 2 additional columns,
\code{Info_group} (cluster id returned by CDHIT) and \code{Info_split} (which allocates
groups to different splits.

If the splitting is impossible (e.g., if the number of clusters is smaller
than the desired number of splits) the function throws a warning and returns
the list object returned by \code{\link[CellaRepertorium:cdhit]{CellaRepertorium::cdhit()}}.
}
\description{
Split the data returned by \code{\link[=extract_labelled_data]{extract_labelled_data()}} into
non-overlapping subsets. Proteins with similarities higher than
a predefined threshold are always placed in the same split to prevent data
leakage. This routine tries to simultaneously approximate
the user-defined proportions, maintain the overall class balance within
each split, and maximise within-splits diversity. See
\code{\link[moses:constructive_heuristic]{moses::constructive_heuristic()}} for details.
}
\author{
Felipe Campelo (\email{f.campelo@aston.ac.uk})
}
