% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxonomy_filter_list.R
\name{taxonomy_filter_list}
\alias{taxonomy_filter_list}
\title{Filter peptides by taxonomy and host}
\usage{
taxonomy_filter_list(
  peptides.list = NULL,
  tax_list = NULL,
  orgIDs = NULL,
  hostIDs = NULL,
  removeIDs = NULL,
  orgID_column = "Info_organism_id",
  hostID_column = "Info_host_id"
)
}
\arguments{
\item{peptides.list}{list object returned by \code{\link[=extract_labelled_data]{extract_labelled_data()}}.
Only one of \code{peptides.df} or \code{peptides.list} should be provided.}

\item{tax_list}{taxonomy list (generated by \code{\link[=get_taxonomy]{get_taxonomy()}})}

\item{orgIDs}{vector of organism taxon IDs to retain.}

\item{hostIDs}{vector of host taxon IDs to retain.}

\item{removeIDs}{vector of organism taxon IDs to remove.}

\item{orgID_column}{name of the column in \code{peptides.df} or \code{peptides.list$df}
containing the source organism IDs}

\item{hostID_column}{name of the column in \code{peptides.df} or \code{peptides.list$df}
containing the host IDs}
}
\value{
data frame with peptide data filtered by the criteria in \strong{orgIDs},
\strong{hostIDs} and \strong{removeIDs}.
}
\description{
Filters a peptides dataframe by source organism and/or host IDs.
}
\details{
Taxonomic filtering works hierarchically - e.g., if a genus id is passed in
\code{orgIDs}, all species under that genus are removed.
\code{removeIDs} is generally used to censor out one or more subgroups under those
listed in \code{orgIDs}. For instance, if we wanted to get data for all
orthopoxvirus (txID:10242) except the Variola virus (txID:10255) we would make
\code{orgIDs = 10242} and \code{removeIDs = 10255}.
}
\author{
Felipe Campelo (\email{fcampelo@gmail.com})
}
