% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_RF.R
\name{fit_RF}
\alias{fit_RF}
\title{fits a single random forest and assess the performance on a holdout split
@importFrom dplyr \%>\%
@importFrom rlang .data}
\usage{
fit_RF(
  df,
  sample.rebalancing,
  holdout.split = NULL,
  threshold,
  ncpus,
  min.node.size,
  ntrees,
  ...
)
}
\description{
fits a single random forest and assess the performance on a holdout split
@importFrom dplyr \%>\%
@importFrom rlang .data
}
