% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_RF.R
\name{fit_RF}
\alias{fit_RF}
\title{Fits a single random forest and assess the performance on a holdout split.}
\usage{
fit_RF(df, sample.rebalancing, holdout.split = NULL, threshold, ncpus, ...)
}
\arguments{
\item{df}{dataframe containing a \code{Class} attribute and one or more feature
attributes (columns starting with "feat_")}

\item{sample.rebalancing}{character: should the model try to compensate class
imbalances during training? See \strong{Dealing with class imbalance} for details.}

\item{holdout.split}{name of split to be used as a holdout set.}

\item{threshold}{probability threshold for attributing a prediction as
\emph{positive}.}

\item{ncpus}{number of cores to use.}

\item{...}{other options to be passed down to \code{\link[ranger:ranger]{ranger::ranger()}}.}
}
\description{
Not to be used separately, this function is called as part of calls to
\code{\link[=fit_model]{fit_model()}}.
}
