% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data_summary.R
\name{get_data_summary}
\alias{get_data_summary}
\title{Extract data summary}
\usage{
get_data_summary(
  peptides.list,
  uid = NULL,
  tax_list = NULL,
  min_pept_grp = 1,
  redundancy_threshold = 0
)
}
\arguments{
\item{peptides.list}{peptides.list object returned by \code{\link[=make_data_splits]{make_data_splits()}}.}

\item{uid}{taxononmy id to extract a subset of the data. \code{NULL} to
summarise the entire dataset}

\item{tax_list}{taxonomy list (generated by \code{\link[=get_taxonomy]{get_taxonomy()}})}

\item{min_pept_grp}{minimum number of peptides in a group to be considered}

\item{redundancy_threshold}{similarity threshold in the \link{0-1} interval for
redundancy removal. If a nonzero value is passed, it uses
\code{\link[moses:extract_clusters_cdhit]{moses::extract_clusters_cdhit()}} for clustering the peptides at the desired
similarity threshold and uses only one representative example per cluster
when calculating the summary statistics.}
}
\value{
1-row data frame with summary information.
}
\description{
Summarise a peptides list and extracts number of peptides, groups and
observations under each class
}
\author{
Felipe Campelo (\email{fcampelo@gmail.com})
}
