% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_data_splits.R
\name{make_data_splits}
\alias{make_data_splits}
\title{Define data splits for predictive modelling.}
\usage{
make_data_splits(
  peptides.list,
  delta,
  w = c(0.5, 0.4, 0.1),
  split_names = NULL,
  split_mode = "rand",
  similarity_threshold = 0.7,
  target_id = NULL,
  tax_list = NULL,
  ncpus = 1,
  seed = NULL,
  cdhit.par.list = list(s = 0.5)
)
}
\arguments{
\item{peptides.list}{list object returned by \code{\link[=extract_labelled_data]{extract_labelled_data()}}, containing
the data frame of windowed epitope data (\strong{$df}), the data frame of
individual peptides (\strong{$peptides}) and the data frame of proteins
(\strong{$proteins}).}

\item{delta}{vector of desired split proportions (must add up to one). It
is useful (but not mandatory) to use a vector that is sorted
in decreasing order.}

\item{w}{vector of weights for function aggregation. Must be of
length 3 and add to 1. All weights must be non-negative. See
package \code{moses} for details.
\code{w[1]}: weight of the objective to obey desired split proportions;
\code{w[2]}: weight of the objective to distribute classes proportionally between splits;
\code{w[3]}: weight of the objective to maximise within-splits diversity.
Notice that \code{w[3]} is ignored if \code{split_mode == "rand"} (in which case
\code{w[1]} and \code{w[2]} are rescaled internally so that \code{w[1] + w[2] = 1}).}

\item{split_names}{optional, vector of names to be given to each split.}

\item{split_mode}{mode of splitting. Use "constructive" for a deterministic
constructive heuristic (calling \code{\link[moses:make_splits_constructive]{moses::make_splits_constructive()}}) or
"rand" for a randomised heuristic (calling \code{\link[moses:make_splits_rand_refine]{moses::make_splits_rand_refine()}}).
Using "rand" is necessary if you plan to use model development strategies in
which some splits need to be redefined, e.g., in repeated cross-validation.}

\item{similarity_threshold}{similarity threshold for grouping observations.
\eqn{1 - similarity_threshold} is passed to \code{\link[CellaRepertorium:cdhit]{CellaRepertorium::cdhit()}} as}

\item{target_id}{vector of lower-level taxonomy ids (integer or character)
that the splitting strategy should try to prioritise
distributing across splits (useful, e.g., if generating splits at a
higher taxonomic level for later re-use at a lower level).}

\item{tax_list}{taxonomy list (generated by \code{\link[=get_taxonomy]{get_taxonomy()}}). Only used if
\code{target_id} is not \code{NULL}.}

\item{ncpus}{positive integer, number of cores to use.}

\item{seed}{seed for pseudorandom number generator. Ignored if
split_mode == "constructive".}

\item{cdhit.par.list}{further parameters to be passed down to
\code{\link[CellaRepertorium:cdhit]{CellaRepertorium::cdhit()}}. Please check that function for details.}
}
\value{
Updated \code{peptides.list} with dataframes containing 2 additional columns,
\code{Info_group} (cluster id returned by CDHIT) and \code{Info_split} (which allocates
groups to different splits. Dataframes \code{df} and \code{peptides} have an extra
two columns indicating whether the observation group contains any observations
from the target taxon (if applicable), and whether each observation's
\code{Info_organism_id} is under the \code{target_id} taxon.

If the splitting is impossible (e.g., if the number of clusters is smaller
than the desired number of splits) the function throws a warning and returns
the list object updated with information about the clustering returned by
\code{\link[CellaRepertorium:cdhit]{CellaRepertorium::cdhit()}}.
}
\description{
Split the data returned by \code{\link[=extract_labelled_data]{extract_labelled_data()}} into
non-overlapping subsets. Proteins with similarities higher than
a predefined threshold are always placed in the same split to prevent data
leakage. See \code{\link[moses:make_splits_constructive]{moses::make_splits_constructive()}} or
\code{\link[moses:make_splits_rand_refine]{moses::make_splits_rand_refine()}} for details.
}
\author{
Felipe Campelo (\email{fcampelo@gmail.com})
}
