% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_IEDB_OrgList.R
\name{process_IEDB_OrgList}
\alias{process_IEDB_OrgList}
\title{Process taxonomic information from IEDB OrganismList XML.}
\usage{
process_IEDB_OrgList(file_path, cl = NULL, save_folder = NULL)
}
\arguments{
\item{file_path}{path (either relative or absolute) to the OrganismList.xml
file retrieved from the IEDB database export (e.g., via \code{\link[=get_IEDB_OrgList]{get_IEDB_OrgList()}}).}

\item{cl}{a SOCK cluster object created using \code{\link[=set_mc_cluster]{set_mc_cluster()}}, or
\code{NULL} if parallel processing is not desired.}

\item{save_folder}{path to folder for saving the output.}
}
\value{
A data frame containing the taxonomic data.
}
\description{
Process taxonomic information from IEDB OrganismList XML.
}
\author{
Felipe Campelo (\email{fcampelo@gmail.com})
}
