% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_peptide_list.R
\name{save_peptide_list}
\alias{save_peptide_list}
\title{Save information from a peptide.list object to files .}
\usage{
save_peptide_list(peptide.list, save_folder)
}
\arguments{
\item{peptide.list}{object of class \code{peptide.list}.}

\item{save_folder}{path to folder for saving the resulting CSV files.}
}
\value{
Vector of file names generated (invisibly).
}
\description{
Saves the full information from objects of type \code{peptide.list} (e.g., output
of functions \code{\link[=extract_peptides]{extract_peptides()}}, \code{\link[=make_data_splits]{make_data_splits()}} or \code{\link[=calc_features]{calc_features()}})
to a set of CSV files for easier sharing and cross-platform use.
}
\author{
Felipe Campelo (\email{f.campelo@aston.ac.uk})
}
