% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_mc_cluster.R
\name{set_mc_cluster}
\alias{set_mc_cluster}
\title{Set multicore cluster}
\usage{
set_mc_cluster(ncpus)
}
\arguments{
\item{ncpus}{number of cores to use}
}
\value{
a SOCK cluster object
}
\description{
This function is a wrapper around \code{\link[parallel:makeCluster]{parallel::makePSOCKcluster()}} followed
by \code{\link[parallel:clusterApply]{parallel::clusterEvalQ()}}, which loads all required packages in the
individual cluster workers.
}
